requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-multipart==0.0.6
pydantic==2.4.2
openai==1.3.3
streamlit==1.28.1
biopython==1.81
pandas==2.1.3
aiofiles==23.2.1
python-dotenv==1.0.0
jinja2==3.1.2 